+++ examples/foxes/sequential/run.py
@@ -1,4 +1,3 @@
-import time
 import argparse
 import matplotlib.pyplot as plt
 import numpy as np